requirements.txt
torch==2.2.0+cu121
transformers>=4.40
tqdm
pandas
openpyxl
pyyaml
spacy>=3.7
spacy[transformers]
negspacy
pytorch-lightning>=2.2
fastapi
uvicorn[standard]